+++ lib/index.ts
@@ -16,3 +16,4 @@ export * from './components/ButtonFeedback';
 export * from './components/ModalWrapper';
 export * from './components/Modal';
 export * from './components/SlideX';
+export * from './components/FloatingActionButton';